backend/requirements.txt
django-filter==22.1
djangorestframework==3.14.0
markdown==3.4.1
